+++ src/main/java/com/ats/tril/model/EnquiryDetail.java
@@ -32,7 +32,7 @@ public class EnquiryDetail {
 	private int itemId;
 	
 	@Column(name = "enq_qty")
-	private int enqQty;
+	private float enqQty;
 	
 	@Column(name = "enq_uom")
 	private String enqUom;
@@ -97,11 +97,11 @@ public class EnquiryDetail {
 		this.itemId = itemId;
 	}
 
-	public int getEnqQty() {
+	public float getEnqQty() {
 		return enqQty;
 	}
 
-	public void setEnqQty(int enqQty) {
+	public void setEnqQty(float enqQty) {
 		this.enqQty = enqQty;
 	}
