+++ src/main/java/com/ats/tril/model/mrn/MrnDetail.java
@@ -20,23 +20,23 @@ public class MrnDetail {
 	
 	private int itemId;
 	
-	private int indentQty;
+	private float indentQty;
 	
-	private int poQty;
+	private float poQty;
 	
-	private int mrnQty;
+	private float mrnQty;
 	
-	private int approveQty;
+	private float approveQty;
 	
-	private int rejectQty;
+	private float rejectQty;
 	
 	private int rejectRemark;
 	
 	private String batchNo;
 	
-	private int issueQty;
+	private float issueQty;
 	
-	private int remainingQty;
+	private float remainingQty;
 	
 	private int poId;
 	
@@ -48,27 +48,27 @@ public class MrnDetail {
 	
 	private int delStatus;
 	
-	private int chalanQty;//added on 13 aug
+	private float chalanQty;//added on 13 aug
 
 
-	public int getChalanQty() {
+	public float getChalanQty() {
 	return chalanQty;
 }
 
-public void setChalanQty(int chalanQty) {
+public void setChalanQty(float chalanQty) {
 	this.chalanQty = chalanQty;
 }
 	@Transient
-	private int mrnQtyBeforeEdit;
+	private float mrnQtyBeforeEdit;
 	
 	
 
 
-	public int getMrnQtyBeforeEdit() {
+	public float getMrnQtyBeforeEdit() {
 		return mrnQtyBeforeEdit;
 	}
 
-	public void setMrnQtyBeforeEdit(int mrnQtyBeforeEdit) {
+	public void setMrnQtyBeforeEdit(float mrnQtyBeforeEdit) {
 		this.mrnQtyBeforeEdit = mrnQtyBeforeEdit;
 	}
 
@@ -80,23 +80,23 @@ public void setChalanQty(int chalanQty) {
 		return itemId;
 	}
 
-	public int getIndentQty() {
+	public float getIndentQty() {
 		return indentQty;
 	}
 
-	public int getPoQty() {
+	public float getPoQty() {
 		return poQty;
 	}
 
-	public int getMrnQty() {
+	public float getMrnQty() {
 		return mrnQty;
 	}
 
-	public int getApproveQty() {
+	public float getApproveQty() {
 		return approveQty;
 	}
 
-	public int getRejectQty() {
+	public float getRejectQty() {
 		return rejectQty;
 	}
 
@@ -108,11 +108,11 @@ public void setChalanQty(int chalanQty) {
 		return batchNo;
 	}
 
-	public int getIssueQty() {
+	public float getIssueQty() {
 		return issueQty;
 	}
 
-	public int getRemainingQty() {
+	public float getRemainingQty() {
 		return remainingQty;
 	}
 
@@ -140,23 +140,23 @@ public void setChalanQty(int chalanQty) {
 		this.itemId = itemId;
 	}
 
-	public void setIndentQty(int indentQty) {
+	public void setIndentQty(float indentQty) {
 		this.indentQty = indentQty;
 	}
 
-	public void setPoQty(int poQty) {
+	public void setPoQty(float poQty) {
 		this.poQty = poQty;
 	}
 
-	public void setMrnQty(int mrnQty) {
+	public void setMrnQty(float mrnQty) {
 		this.mrnQty = mrnQty;
 	}
 
-	public void setApproveQty(int approveQty) {
+	public void setApproveQty(float approveQty) {
 		this.approveQty = approveQty;
 	}
 
-	public void setRejectQty(int rejectQty) {
+	public void setRejectQty(float rejectQty) {
 		this.rejectQty = rejectQty;
 	}
 
@@ -168,11 +168,11 @@ public void setChalanQty(int chalanQty) {
 		this.batchNo = batchNo;
 	}
 
-	public void setIssueQty(int issueQty) {
+	public void setIssueQty(float issueQty) {
 		this.issueQty = issueQty;
 	}
 
-	public void setRemainingQty(int remainingQty) {
+	public void setRemainingQty(float remainingQty) {
 		this.remainingQty = remainingQty;
 	}
